benchmarks/Text2SpeechCensoringWorkflow/Text2SpeechCensoringWorkflow_Text2Speech/requirements.txt
# Function dependencies, for example:
# package>=version
gTTS==2.2.3
google-cloud-storage==1.43.0
flask==1.1.2
google-cloud-datastore==1.13.2
google-cloud-pubsub>=0.28.1
